% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gscatm-package}
\alias{gscatm}
\alias{gscatm-package}
\title{gscatm: gscatm}
\description{
Integration of Generalized Structured Component Analysis (GSCA) with topic modeling to allow topics to vary across covariates.
}
\keyword{internal}
