% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_topics.R
\name{predict_topics}
\alias{predict_topics}
\title{Predict Topic Proportions}
\usage{
predict_topics(model, new_dtm, new_covariates)
}
\arguments{
\item{model}{Fitted GSCA topic model}

\item{new_dtm}{New document-term matrix}

\item{new_covariates}{New covariates}
}
\value{
Matrix of predicted topic proportions
}
\description{
Predict Topic Proportions
}
