% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscatm_plots.R
\name{print_topic_summary}
\alias{print_topic_summary}
\title{Print Topic Summary}
\usage{
print_topic_summary(model, n_words = 10, n_topics = NULL)
}
\arguments{
\item{model}{Fitted GSCA topic model object}

\item{n_words}{Number of top words to show per topic}

\item{n_topics}{Number of topics to show (NULL for all)}
}
\description{
Prints a formatted summary of topics with their top words and proportions
}
