% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_topic_model.R
\name{search_optimal_topics}
\alias{search_optimal_topics}
\title{Search for Optimal Number of Topics}
\usage{
search_optimal_topics(dtm, covariates, K_seq, n = 5, ...)
}
\arguments{
\item{dtm}{Document-term matrix (dgCMatrix)}

\item{covariates}{Data frame or matrix of covariates}

\item{K_seq}{Sequence of topic numbers to evaluate}

\item{n}{Number of top K values to return}

\item{...}{Additional arguments to pass to fit_gsca_topic_model}
}
\value{
A list containing:
\describe{
\item{top_K}{Top n K values based on Semantic Coherence}
\item{metrics}{A list of computed metrics for each K}
\item{models}{List of fitted models for the top n K values}
}
}
\description{
This function searches for the optimal number of topics by fitting the GSCA
Topic Model across a range of topic numbers and computing model diagnostics.
}
